+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 **New decoders:**
 
-- `select()` (see [docs](https://decoders.cc/api.html#select))
 - `bigint` (see [docs](https://decoders.cc/api.html#bigint))
+- `decimal` (see [docs](https://decoders.cc/api.html#decimal))
+- `hexadecimal` (see [docs](https://decoders.cc/api.html#hexadecimal))
+- `numeric` (see [docs](https://decoders.cc/api.html#numeric))
+- `select()` (see [docs](https://decoders.cc/api.html#select))
+
+**Renamed decoders:**
+
 - `nullish()` (see [docs](https://decoders.cc/api.html#nullish), replaces `maybe()`)
 
 **New features:**